model_training/requirements.txt
ultralytics
opencv-python
torch
torchvision
streamlit
Pillow
